+++ .compressor.ps1
@@ -54,6 +54,6 @@ if (Test-Path $temp) {
 
 New-Item -Path $temp -ItemType Directory | Out-Null
 Copy-Item $foldersToInclude -Destination $temp -Exclude $filesToExclude -Recurse
-Get-ChildItem $temp -Recurse | Where-Object { $_.PSIsContainer -and $_.Name -cin $foldersToRemove } | Remove-Item -Recurse -Force
+Get-ChildItem "$temp\$name\embeds\" -Recurse | Where-Object { $_.PSIsContainer -and $_.Name -cin $foldersToRemove } | Remove-Item -Recurse -Force
 7z a -tzip -mx9 "..\$name-$version.zip" (Get-ChildItem $temp)
 Remove-Item $temp -Recurse -Force