+++ ls_Monobrow/init.lua
@@ -79,6 +79,7 @@ addon:RegisterEvent("ADDON_LOADED", function(arg)
 
 	addon:RegisterEvent("PLAYER_LOGIN", function()
 		addon.Font:Update()
+
 		-- to fetch and cache the tracked house data
 		local guid = C_Housing.GetTrackedHouseGuid()
 		if guid then
@@ -89,16 +90,17 @@ addon:RegisterEvent("ADDON_LOADED", function(arg)
 		-- this way I'm able to show honour and reputation bars simultaneously, in the default UI enabling honour tracking
 		-- resets faction tracking
 		local isHonorBarHooked = false
-		local function hookHonor()
+
+		EventUtil.ContinueOnAddOnLoaded("Blizzard_PVPUI", function()
 			if not isHonorBarHooked then
-				for _, panel in next, {"CasualPanel", "TrainingGroundsPanel"} do
+				for _, panel in next, {"CasualPanel", "RatedPanel", "TrainingGroundsPanel"} do
 					PVPQueueFrame.HonorInset[panel].HonorLevelDisplay:SetScript("OnMouseUp", function()
 						if IsShiftKeyDown() then
 							if IsWatchingHonorAsXP() then
-								PlaySound(857) -- SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF
+								PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF)
 								SetWatchingHonorAsXP(false)
 							else
-								PlaySound(856) -- SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON
+								PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
 								SetWatchingHonorAsXP(true)
 							end
 
@@ -119,20 +121,71 @@ addon:RegisterEvent("ADDON_LOADED", function(arg)
 
 				isHonorBarHooked = true
 			end
-		end
-
-		EventUtil.ContinueOnAddOnLoaded("Blizzard_PVPUI", hookHonor)
+		end)
 
 		ReputationFrame.ReputationDetailFrame.WatchFactionCheckbox:SetScript("OnClick", function(self)
 			if self:GetChecked() then
-				PlaySound(856) -- SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON
+				PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
 				C_Reputation.SetWatchedFactionByIndex(C_Reputation.GetSelectedFaction())
 			else
-				PlaySound(857) -- SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF
+				PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF)
 				C_Reputation.SetWatchedFactionByIndex(0)
 			end
 
 			bar:UpdateSegments()
+		end)
+
+		local function createWatchButton(parent)
+			local button = CreateFrame("CheckButton", "$parentWatchButton", parent, "UICheckButtonArtTemplate")
+			button:SetSize(21, 21)
+
+			local label = button:CreateFontString(nil, "ARTWORK", "GameFontHighlightSmall")
+			label:SetPoint("RIGHT", button, "LEFT", 0, 0)
+			label:SetText(_G.MAJOR_FACTION_WATCH_FACTION_BUTTON_LABEL)
+			label:SetJustifyH("LEFT")
+			button.Label = label
+
+			return button
+		end
+
+		EventUtil.ContinueOnAddOnLoaded("Blizzard_EncounterJournal", function()
+			local button = createWatchButton(EncounterJournal.MonthlyActivitiesFrame.ThresholdContainer)
+			button:SetPoint("BOTTOMRIGHT", EncounterJournal.MonthlyActivitiesFrame.ThresholdContainer.ThresholdBar, "TOPRIGHT", 0, 0)
+			button:SetScript("OnShow", function(self)
+				self:SetChecked(C.db.char.travel_points)
+			end)
+			button:SetScript("OnClick", function(self)
+				C.db.char.travel_points = not C.db.char.travel_points
+				self:GetChecked(C.db.char.travel_points)
+
+				if self:GetChecked() then
+					PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
+				else
+					PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF)
+				end
+
+				bar:UpdateSegments()
+			end)
+		end)
+
+		EventUtil.ContinueOnAddOnLoaded("Blizzard_HousingDashboard", function()
+			local button = createWatchButton(HousingDashboardFrame.HouseInfoContent.ContentFrame.InitiativesFrame.InitiativeSetFrame)
+			button:SetPoint("BOTTOMRIGHT", HousingDashboardFrame.HouseInfoContent.ContentFrame.InitiativesFrame.InitiativeSetFrame.ProgressBar, "TOPRIGHT", 0, 0)
+			button:SetScript("OnShow", function(self)
+				self:SetChecked(C.db.char.endeavor)
+			end)
+			button:SetScript("OnClick", function(self)
+				C.db.char.endeavor = not C.db.char.endeavor
+				self:GetChecked(C.db.char.endeavor)
+
+				if self:GetChecked() then
+					PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
+				else
+					PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF)
+				end
+
+				bar:UpdateSegments()
+			end)
 		end)
 	end)
 end)